refactor: constrain room update patches to observed fields
